+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `randomcontigs` library support for generating random FASTA contigs.
+- Add `faqt make-random-contigs`, which writes random FASTA contigs to stdout by default or to `-o/--output`.
+
 ## [0.1.1] - 2026-04-10
 
 ### Changed